+++ packages/blocks/src/components/index.ts
@@ -1,7 +1,6 @@
 export * from './block-hub.js';
 export * from './drag-handle.js';
 export * from './file-drop-manager.js';
-export { FormatQuickBar } from './format-quick-bar/index.js';
 export * from './import-page/index.js';
 export * from './menu-divider.js';
 export * from './remote-selection/remote-selection.js';
