+++ packages/blocks/src/page-block/default/selection-manager/selection-manager.ts
@@ -39,9 +39,11 @@ import {
   Rect,
 } from '../../../__internal__/index.js';
 import { activeEditorManager } from '../../../__internal__/utils/active-editor-manager.js';
-import { showFormatQuickBar } from '../../../components/format-quick-bar/index.js';
+import {
+  showFormatQuickBar,
+  showFormatQuickBarByClicks,
+} from '../../../components/format-quick-bar/index.js';
 import type { NoteBlockModel } from '../../../note-block/note-model.js';
-import { showFormatQuickBarByClicks } from '../../index.js';
 import {
   calcCurrentSelectionPosition,
   getNativeSelectionMouseDragInfo,
